Don't install when uploading
This commit fixes #1037 by removing the Berksfile#install procedure
from the Berksfile#upload method. Instead, the uploader now relies
entirely on the lockfile and universe to upload dependencies.

The airity of Berksfile#upload is now changed to accept the list of
cookbook names as strings, or an options hash, or both.